renamed files to adhere to next js standards
