+++ nikhilyates-two/src/app/portfolio/components/leftMain/LeftMain.tsx
