+++ nikhilyates-two/src/app/portfolio/components/rightMain/RightMain.tsx
